create zip archive in temp dir
ok so there are some problems when you're creating the archive but
you're at the root of the project. Namely, the mod package tries to put
the zip inside of itself. So we have to go around that by creating the
zip somewhere else and moving it. This is the first step in that
process.