move temp file into place (broken)
this doens't work if os.TempDir and the current directory are in
different filesystems because it's a cross-filesystem move. Annoying af
honestly. I think I have to use zip.Create instead of zip.CreateFromDir
and handle the file exclusion myself.

I could in theory create the zip into memory and then write it to avoid
CreateFromDir trying to circularly zip up its own output but I don't
like that.