ok, stepping into an actual structure now
proved that I could manage the protocol with something that was faked,
starting to try to serve something for real. The idea is that you would
have a directory that contains all of your module zips, and the mir
server would serve modules from that directory, and it would query that
directory to find out which versions were available and which were
latest.

I'm not sure about the version list and retractions. It seems like if a
module is retracted you still want it to be in the list of available
modules, and that a retraction is not signaled by the absence of that
version in the list endpoint, but by the module file of the latest
version itself.

    /srv/mir
    └── modules
        └── orel.li
            └── mir@v0.0.0-pre1.zip